+++ CJBCheatsMenu/ModEntry.cs
@@ -1,5 +1,6 @@
 using System.Linq;
 using CJBCheatsMenu.Framework;
+using CJBCheatsMenu.Framework.Models;
 using StardewModdingAPI;
 using StardewModdingAPI.Events;
 using StardewValley;
@@ -18,6 +19,9 @@ namespace CJBCheatsMenu
         /// <summary>The mod settings.</summary>
         private ModConfig Config;
 
+        /// <summary>The warps to show in the menu.</summary>
+        private ModDataWarp[] Warps;
+
         /// <summary>The cheats helper.</summary>
         private Cheats Cheats;
 
@@ -29,11 +33,20 @@ namespace CJBCheatsMenu
         /// <param name="helper">Provides simplified APIs for writing mods.</param>
         public override void Entry(IModHelper helper)
         {
+            // load config/data
             this.Config = helper.ReadConfig<ModConfig>();
+            this.Warps = helper.Data.ReadJsonFile<ModDataWarp[]>("data/warps.json");
+            if (this.Warps == null)
+            {
+                this.Monitor.Log("Some of the mod files are missing (data/warps.json); try reinstalling this mod.", LogLevel.Error);
+                return;
+            }
             this.Monitor.Log($"Started with menu key {this.Config.OpenMenuKey}.", LogLevel.Trace);
 
+            // load cheats
             this.Cheats = new Cheats(this.Config);
 
+            // hook events
             helper.Events.Display.Rendered += this.OnRendered;
             helper.Events.Display.MenuChanged += this.OnMenuChanged;
 
@@ -77,7 +90,7 @@ namespace CJBCheatsMenu
 
             // open menu
             if (e.Button == this.Config.OpenMenuKey)
-                Game1.activeClickableMenu = new CheatsMenu(this.Config.DefaultTab, this.Config, this.Cheats, this.Helper.Translation, this.Monitor);
+                Game1.activeClickableMenu = new CheatsMenu(this.Config.DefaultTab, this.Config, this.Warps, this.Cheats, this.Helper.Translation, this.Monitor);
 
             // handle button if applicable
             else