+++ CJBCheatsMenu/release-notes.md
@@ -4,6 +4,8 @@
 ## Upcoming release
 * Updated for Stardew Valley 1.4.
 * Added support for holding the grow crop/tree keys while moving the cursor, so you can easily grow larger fields.
+* Warps are now sorted alphabetically.
+* Warps can now be customised by editing `data/warps.json`.
 * Updated translations. Thanks to jahangmar (updated German), overwritten (updated Korean), qqkookie (updated Korean), Riazaia (updated Spanish), shiro2579 (updated Portuguese), and shirutan (updated Japanese)!
 
 ## 1.20.1